+++ zcash_client_backend/src/tor/http/cryptex/gate_io.rs
@@ -2,7 +2,7 @@ use hyper::StatusCode;
 use rust_decimal::Decimal;
 use serde::Deserialize;
 
-use super::{Exchange, ExchangeData, RETRY_LIMIT};
+use super::{retry_filter, Exchange, ExchangeData, RETRY_LIMIT};
 use crate::tor::{Client, Error};
 
 /// Querier for the Gate.io exchange.
@@ -36,11 +36,12 @@ impl Exchange for GateIo {
         // API documentation:
         // https://www.gate.io/docs/developers/apiv4/#retrieve-ticker-information
         let res = client
-            .get_json::<Vec<GateIoData>>(
+            .http_get_json::<Vec<GateIoData>>(
                 "https://api.gateio.ws/api/v4/spot/tickers?currency_pair=ZEC_USDT"
                     .parse()
                     .unwrap(),
                 RETRY_LIMIT,
+                retry_filter,
             )
             .await?;
         let data = res.into_body().into_iter().next().ok_or(Error::Http(